requirements.txt
# Core Fireworks dependencies
fireworks-ai
langchain-fireworks

# Image processing
Pillow

# Utilities
python-dotenv
numpy